#205; changes pipeline state back to not use latest route.

+++ execute/step/scripts/Ubuntu_16.04/templates/download.sh
@@ -12,6 +12,8 @@ export RUN_ARTIFACT_NAME="%%runArtifactName%%"
 export RUN_WORKSPACE_DIR="%%runWorkspaceDir%%"
 export PIPELINE_ARTIFACT_URL="%%pipelineArtifactUrl%%"
 export PIPELINE_ARTIFACT_URL_OPTS="%%pipelineArtifactUrlOpts%%"
+export PIPELINE_ARTIFACT_HEAD_URL="%%pipelineArtifactHeadUrl%%"
+export PIPELINE_ARTIFACT_HEAD_URL_OPTS="%%pipelineArtifactHeadUrlOpts%%"
 export PIPELINE_ARTIFACT_NAME="%%pipelineArtifactName%%"
 export PIPELINE_WORKSPACE_DIR="%%pipelineWorkspaceDir%%"
 
@@ -127,40 +129,68 @@ download_pipeline_state() {
     rm -rf $PIPELINE_WORKSPACE_DIR/*
   fi
 
-  echo 'Downloading pipeline state'
-
-  local download_cmd="curl \
+  local check_artifact_cmd="curl \
       -s \
       --connect-timeout 60 \
       --max-time 120 \
-      -XGET '$PIPELINE_ARTIFACT_URL' \
-      -o $archive_file"
+      -o /dev/null \
+      -w \"%{http_code}\" \
+      --head '$PIPELINE_ARTIFACT_HEAD_URL'"
 
-  if [ ! -z "$PIPELINE_ARTIFACT_URL_OPTS" ]; then
-    download_cmd="curl \
+  if [ ! -z "$PIPELINE_ARTIFACT_HEAD_URL_OPTS" ]; then
+    check_artifact_cmd="curl \
       -s \
       --connect-timeout 60 \
       --max-time 120 \
-      $PIPELINE_ARTIFACT_URL_OPTS \
-      -XGET '$PIPELINE_ARTIFACT_URL' \
-      -o $archive_file"
+      -o /dev/null \
+      -w \"%{http_code}\" \
+      $PIPELINE_ARTIFACT_HEAD_URL_OPTS \
+      --head '$PIPELINE_ARTIFACT_HEAD_URL'"
   fi
 
-  eval "$download_cmd"
+  echo "Executing: $check_artifact_cmd"
 
-  tar -xzf $archive_file -C $PIPELINE_WORKSPACE_DIR
-  rm $archive_file
-  echo 'Downloaded pipeline state'
+  local check_artifact=$(eval $check_artifact_cmd)
 
-  if [ -z "$(ls -A $PIPELINE_WORKSPACE_DIR)" ]; then
-    echo "Pipeline state is empty."
+  if [ $check_artifact -eq 200 ]; then
+    echo 'Downloading pipeline state'
+
+    local download_cmd="curl \
+        -s \
+        --connect-timeout 60 \
+        --max-time 120 \
+        -XGET '$PIPELINE_ARTIFACT_URL' \
+        -o $archive_file"
+
+    if [ ! -z "$PIPELINE_ARTIFACT_URL_OPTS" ]; then
+      download_cmd="curl \
+        -s \
+        --connect-timeout 60 \
+        --max-time 120 \
+        $PIPELINE_ARTIFACT_URL_OPTS \
+        -XGET '$PIPELINE_ARTIFACT_URL' \
+        -o $archive_file"
+    fi
+
+    eval "$download_cmd"
+
+    tar -xzf $archive_file -C $PIPELINE_WORKSPACE_DIR
+    rm $archive_file
+    echo 'Downloaded pipeline state'
+
+    if [ -z "$(ls -A $PIPELINE_WORKSPACE_DIR)" ]; then
+      echo "Pipeline state is empty."
+    else
+      for file in $PIPELINE_WORKSPACE_DIR/*; do
+        md5sum $file
+      done > $STEP_WORKSPACE_DIR/checksums.txt
+
+      cat $STEP_WORKSPACE_DIR/checksums.txt | sort > $STEP_WORKSPACE_DIR/pipelineStateChecksums.txt
+      rm $STEP_WORKSPACE_DIR/checksums.txt
+    fi
+
   else
-    for file in $PIPELINE_WORKSPACE_DIR/*; do
-      md5sum $file
-    done > $STEP_WORKSPACE_DIR/checksums.txt
-
-    cat $STEP_WORKSPACE_DIR/checksums.txt | sort > $STEP_WORKSPACE_DIR/pipelineStateChecksums.txt
-    rm $STEP_WORKSPACE_DIR/checksums.txt
+    echo 'No previous pipeline state'
   fi
 }
